Make Sequelize usage more consistent with other origin packages
