+++ origin-token-transfer/server/seeders/20181018104132-demo-grants.js
@@ -3,50 +3,50 @@
 // TODO: create seeder that reads in a CSV file
 module.exports = {
   up: (queryInterface) => {
-    return queryInterface.bulkInsert('Grants', [
+    return queryInterface.bulkInsert('Grant', [
       {
         email: 'cuong@originprotocol.com',
-        grantedAt: '2010-01-31 00:00:00',
+        granted_at: '2010-01-31 00:00:00',
         amount: 120000,
-        totalMonths: 48,
-        cliffMonths: 12,
+        total_months: 48,
+        cliff_months: 12,
         vested: 120000,
         transferred: 100000,
-        createdAt: new Date(),
-        updatedAt: new Date()
+        created_at: new Date(),
+        updated_at: new Date()
       },
       {
         email: 'cuong@originprotocol.com',
-        grantedAt: '2017-01-31 00:00:00',
+        granted_at: '2017-01-31 00:00:00',
         amount: 480,
-        totalMonths: 48,
-        cliffMonths: 12,
+        total_months: 48,
+        cliff_months: 12,
         vested: 0,
         transferred: 0,
-        createdAt: new Date(),
-        updatedAt: new Date()
+        created_at: new Date(),
+        updated_at: new Date()
       },
       {
         email: 'cuong@originprotocol.com',
-        grantedAt: '2017-08-01 00:00:00',
+        granted_at: '2017-08-01 00:00:00',
         amount: 112,
-        totalMonths: 48,
-        cliffMonths: 12,
+        total_months: 48,
+        cliff_months: 12,
         vested: 0,
         transferred: 0,
-        createdAt: new Date(),
-        updatedAt: new Date()
+        created_at: new Date(),
+        updated_at: new Date()
       },
       {
         email: 'cuong@originprotocol.com',
-        grantedAt: '2018-08-01 00:00:00',
+        granted_at: '2018-08-01 00:00:00',
         amount: 120,
-        totalMonths: 48,
-        cliffMonths: 12,
+        total_months: 48,
+        cliff_months: 12,
         vested: 0,
         transferred: 0,
-        createdAt: new Date(),
-        updatedAt: new Date()
+        created_at: new Date(),
+        updated_at: new Date()
       }
     ], {})
   },
