+++ origin-token-transfer/server/src/models/index.js
@@ -8,6 +8,8 @@ const env = process.env.NODE_ENV || 'development'
 const config = require('../../config/config.js')[env]
 const db = {}
 
+console.log(config)
+
 let sequelize
 if (config.use_env_variable) {
   sequelize = new Sequelize(process.env[config.use_env_variable], config)
